Huge refactor of reducers
