+++ src/client/vzReducer/setIsSettingsOpenReducer.ts
@@ -1,0 +1,9 @@
+import { VZAction, VZState } from '.';
+
+export const setIsSettingsOpenReducer = (
+  state: VZState,
+  action: VZAction,
+): VZState =>
+  action.type === 'set_is_settings_open'
+    ? { ...state, isSettingsOpen: action.value }
+    : state;
