+++ src/client/vzReducer/setThemeReducer.ts
@@ -1,0 +1,9 @@
+import { VZAction, VZState } from '.';
+
+export const setThemeReducer = (
+  state: VZState,
+  action: VZAction,
+): VZState =>
+  action.type === 'set_theme'
+    ? { ...state, theme: action.themeLabel }
+    : state;
